Create BaseTrustPage
As all of the trust pages share the same trust data, and many of the same methods it makes sense to create a base Page Object Model for this section of the application.
